postgres: use big serial vs serial to match sqlite

+++ postgres/migrations/20200813125704_create_tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS teams (
-    id          SERIAL PRIMARY KEY,
+    id          BIGSERIAL PRIMARY KEY,
     name        TEXT NOT NULL 
 );
 
@@ -16,7 +16,7 @@ CREATE UNIQUE INDEX IF NOT EXISTS
 
 CREATE TABLE IF NOT EXISTS members (
     user_id     TEXT NOT NULL,
-    team_id     INTEGER NOT NULL,
+    team_id     BIGINT NOT NULL,
     FOREIGN KEY(user_id) REFERENCES users(id),
     FOREIGN KEY(team_id) REFERENCES teams(id),
     UNIQUE(user_id, team_id)
